requirements.txt
amqp==1.4.9
anyjson==0.3.3
APScheduler==3.5.3
beautifulsoup4==4.6.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.8.24
chardet==3.0.4
DBUtils==1.3
Django==1.10
django-celery==3.2.2
django-cors-headers==2.4.0
django-filter==1.1.0
djangorestframework==3.8.2
funcsigs==1.0.2
futures==3.2.0
gevent==1.3.7
greenlet==0.4.15
idna==2.7
image==1.5.27
jpush==3.3.0
kombu==3.0.37
lxml==4.2.5
Markdown==3.0.1
Pillow==5.3.0
pycryptodome==3.6.6
pytz==2018.5
qiniu==7.2.2
qrcode==6.0
redis==2.10.6
requests==2.19.1
six==1.11.0
tzlocal==1.5.1
urllib3==1.23
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0
pyDes==2.0.1